Added progress markers for network calls (start/end/cancel) -with the possibility to add intermediate markers (e.g. when fully updating a project, add a mark for each updated issue).
UnbinderHost is no longer used, but still there in case it's
going to be needed in the future.
Creating a new adapter whenever a parent entity was selected
and 'killing' the existing one when it was no longer needed
was generating problems, so the adapters are now created only
once, and re-used. (click on the same project to collapse and
expand it in a very quick succession would crash the app with
a NullPointerException inside the recycler view)